pyproject.toml
[tool.poetry]
name = "quotapilot"
version = "0.1.0"
description = "QuotaPilot - free-tier-first LLM router (skeleton)"
authors = ["QuotaPilot contributors <dev@quotapilot.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic = "^2.7.0"
httpx = "^0.27.0"
motor = "^3.4.0"
uvicorn = "^0.29.0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

